requirements.txt
django==6.0.4
psycopg2-binary==2.9.11
pillow==12.2.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.2.0